+++ NEWS.md
@@ -1,3 +1,8 @@
+# quickNmix 1.0.4
+
+* Updated parallel computing capabilities by adding fitNmixPara (uses optimParallel)
+* Added new functions for FFT N-mixture computing: pCountOpenFFT
+
 # quickNmix 1.0.3
 
 * CRAN re-submission with test example added for fitNmix and for nll.